Add timepoints and registers
This is a major revision. Basic version of timepoints has been
added. This allows for better specification of circuits. A function
to draw circuits using text has been implemented.

Another addition is registers. These changes have not been
incorporated into the Circuit class as yet, as this will require
significant revision to the timepoint code.